requirements.txt
fastapi
uvicorn[standard]
torch
transformers
pydantic
google-cloud-storage
enum34